refactor (ui): revamp tier component
