+++ openmetadata-ui/src/main/resources/ui/src/components/common/TierCard/TierCard.interface.ts
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2022 Collate.
+ *  Copyright 2023 Collate.
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
  *  You may obtain a copy of the License at
@@ -10,9 +10,9 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-
-import { LoadingState } from 'Models';
-import { HTMLAttributes } from 'react';
+import { EntityReference } from 'generated/type/entityReference';
+import { TableDetail } from 'Models';
+import { ReactNode } from 'react';
 
 export type CardWithListItems = {
   id: string;
@@ -21,13 +21,14 @@ export type CardWithListItems = {
   title: string;
 };
 
-export interface Props extends HTMLAttributes<HTMLDivElement> {
-  index: number;
-  card: CardWithListItems;
-  isActive: boolean;
-  isSelected: boolean;
-  tierStatus: LoadingState;
-  onSave: (updatedTier: string) => void;
-  onCardSelect: (cardId: string) => void;
-  onRemove?: () => void;
+export interface TierCardProps {
+  currentTier?: string;
+  updateTier?: (value: string) => void;
+  onSave?: (
+    owner?: EntityReference,
+    tier?: TableDetail['tier'],
+    isJoinable?: boolean
+  ) => Promise<void>;
+  removeTier?: () => void;
+  children?: ReactNode;
 }
