+++ openmetadata-ui/src/main/resources/ui/src/styles/app.less
@@ -65,6 +65,9 @@
     color: @primary;
   }
 }
+.text-color-inherit {
+  color: inherit;
+}
 .error-text {
   color: #ff4c3b;
 }
